Restore thread-safety to the integration test suite w.r.t. `env::set_var`.
Serialize all tests that use `std::env::set_var` & isolate them. See the
comments in key_log_file_env.rs for details.

Also add notes about the fact that these tests aren't really testing the
functionality.

Use a whitespace-smart diff tool to compare the new file to what was in
api.rs:
```
git difftool HEAD^1:rustls/tests/api.rs rustls/tests/key_log_file_env.rs
```
